Extracts parameter generation into own file
